+++ CHANGELOG.rst
@@ -4,9 +4,11 @@ CHANGELOG
 6.0.5 (unreleased)
 ------------------
 
+- Fix register schema
+  [bloodbare]
+
 - Fix async test without pytest mark
   [masipcat]
-
 
 6.0.4 (2020-07-29)
 ------------------